+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 <!-- markdownlint-disable MD024 -->
 <!-- markdownlint-disable MD041 -->
 
+## [1.0.1](https://github.com/Yegorich555/webpack-mock-server/compare/v0.0.1...v1.0.1) (coming soon)
+
+* fix '__dirname under wepback is not absolute`
+
 ## [1.0.0](https://github.com/Yegorich555/webpack-mock-server/compare/v0.0.4...v1.0.0) (February 7, 2020)
 
 * **BREAKING CHANGE**: improve export for NodeJs-require (use **require('...')** instead of **require('...').default**)